MCU_43 - Separated out an abstract AuxiliaryUpgradeConsole class for public API use - Optimized the enumeration of upgrade slots by reusing arrays instead of creating new objects
